Add <BackgroundPreview> placeholder component as child of  <App>
Because:
- To allow unit testing
- To aid visual layout

<BackgroundPreview> will manage an array of <canvas>
elements, each of which has a single frame of video
drawn into it. The component receives a redux state
value, 'backgroundFrame', that is an index into the
array of <canvas> elements. At this point, the component
has placeholder displays text that reflects the current value
of 'backgroundFrame' for the purpuses of unit testing.