Add <BackgroundPreview> component and supporting utility
Because:
- Enables user to preview their <BackgroundSelector> designated  image
- Utility gives developer asynch function to gather video metadata

The <BackgroundPreview> component is passed a video file path
along with the video's metadata (width, height, duration) and
the number of frames to extract from the video (backgroundFramesMax).
<CanvasVideoFrameCollection> creates an array of <CanvasVideoFrame>
components which render a <canvas> that displays a 2d context-enabled
frame of video. When the redux state value 'backgroundFrame' is updated
by <BackgroundSelector>, the value is updated in <CanvasVideoFramesCollection>
and the corresponding <CanvasVideoFrame> is shown.